+++ federatedscope/core/trainers/utils.py
@@ -30,3 +30,21 @@ def filter_by_specified_keywords(param_name, filter_keywords):
             preserve = False
             break
     return preserve
+
+
+def move_to(obj, device):
+    import torch
+    if torch.is_tensor(obj):
+        return obj.to(device)
+    elif isinstance(obj, dict):
+        res = {}
+        for k, v in obj.items():
+            res[k] = move_to(v, device)
+        return res
+    elif isinstance(obj, list):
+        res = []
+        for v in obj:
+            res.append(move_to(v, device))
+        return res
+    else:
+        raise TypeError("Invalid type for move_to")
